% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_from_repo}
\alias{download_from_repo}
\title{Download Data from Repository}
\usage{
download_from_repo(format, dest_dir, overwrite, repo_url, github_token)
}
\arguments{
\item{format}{Character string specifying the data format to download, either "duckdb" or "csv".}

\item{dest_dir}{Character string specifying the destination directory for downloaded files.}

\item{overwrite}{Logical indicating whether to overwrite existing files.}

\item{repo_url}{Character string specifying the repository URL.}

\item{github_token}{Character string containing a GitHub Personal Access Token (PAT).}
}
\value{
Character string with the path to the downloaded data.
}
\description{
Downloads plant water potential data from a repository.
}
\keyword{internal}
