% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_from_zenodo}
\alias{download_from_zenodo}
\title{Download Data from Zenodo}
\usage{
download_from_zenodo(format, dest_dir, overwrite, zenodo_doi)
}
\arguments{
\item{format}{Character string specifying the data format to download, either "duckdb" or "csv".}

\item{dest_dir}{Character string specifying the destination directory for downloaded files.}

\item{overwrite}{Logical indicating whether to overwrite existing files.}

\item{zenodo_doi}{Character string specifying the Zenodo DOI.}
}
\value{
Character string with the path to the downloaded data.
}
\description{
Downloads plant water potential data from Zenodo.
}
\keyword{internal}
