% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSInetR-package.R
\docType{package}
\name{PSInetR-package}
\alias{PSInetR}
\alias{PSInetR-package}
\title{PSInetR: Access Plant Water Potential Data}
\description{
Provides functions to access and work with plant water potential data from various sources. Allows downloading data from repositories or Zenodo, and includes utilities for data validation and manipulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://PSInetRCN.github.io/PSInetR/}
  \item Report bugs at \url{https://github.com/PSInetRCN/PSInetR/issues}
}

}
\author{
\strong{Maintainer}: Keaton Wilson \email{keatonwilson@me.com}

}
\keyword{internal}
