% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Get Default Data Directory}
\usage{
get_data_dir()
}
\value{
Character string with the path to the default data directory.
}
\description{
Returns the default directory for storing downloaded data.
This directory is the current working directory.
}
\keyword{internal}
