% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSINetR-package.R
\docType{package}
\name{PSINetR-package}
\alias{PSINetR}
\alias{PSINetR-package}
\title{PSINetR: Access Plant Water Potential Data}
\description{
Provides functions to access and work with plant water potential data from various sources. Allows downloading data from repositories or Zenodo, and includes utilities for data validation and manipulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://PSInetRCN.github.io/PSINetR/}
  \item Report bugs at \url{https://github.com/PSInetRCN/PSINetR/issues}
}

}
\author{
\strong{Maintainer}: Keaton Wilson \email{keatonwilson@me.com}

}
\keyword{internal}
